Make DSU generic and add value merging support
Refactored the DSU implementation to be generic over a value type supporting Add and Default traits. Added value storage and merging logic to DSU, and updated TypeResolver to use the new generic DSU with a Node struct that supports addition. This enables tracking and merging of associated data for each set in the DSU.